requirements.txt
asgiref==3.8.1
chardet==5.2.0
cssselect2==0.7.0
Django==5.1.1
django-filer==3.2.3
django-libs==2.0.3
django-polymorphic==3.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
easy-thumbnails==2.10
inflection==0.5.1
lxml==5.3.0
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
uritemplate==4.1.1
webencodings==0.5.1
